refactor: LottoResult 클래스의 역할 분리 (LottoMatchCalculator, LottoStatisticResult)
