+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -1,11 +1,11 @@
 package lotto.controller
 
 import lotto.model.Lotto
-import lotto.model.LottoResult
+import lotto.model.LottoMatchCalculator
+import lotto.model.LottoStatisticResult
 import lotto.model.LottoStore
 import lotto.model.LottoTicketCounter
 import lotto.model.RandomLottoGenerator
-import lotto.model.Rank
 import lotto.model.WinningLotto
 import lotto.view.InputView
 import lotto.view.OutputView
@@ -24,9 +24,11 @@ class LottoController {
         outputView.printLottoBundle(lottoBundle)
 
         val winningLotto = WinningLotto(getWinningLotto(), getBonusNumber())
-        val winningCounts = calculateResult(lottoBundle, winningLotto)
-        val profit = calculateProfit(purchase, winningCounts)
-        outputView.printResult(winningCounts, profit)
+        val lottoMatchResult = LottoMatchCalculator().getWinningCounts(lottoBundle, winningLotto)
+
+        val lottoStatisticResult = LottoStatisticResult().calculateProfit(purchase.toInt(), lottoMatchResult)
+
+        outputView.printResult(lottoMatchResult, lottoStatisticResult)
     }
 
     private fun getWinningLotto(): Lotto = Lotto(inputView.inputWinningNumbers())
@@ -35,19 +37,5 @@ class LottoController {
 
     private fun purchaseLottos(count: Int): List<Lotto> {
         return lottoStore.getTickets(count, RandomLottoGenerator())
-    }
-
-    private fun calculateResult(
-        lottoBundle: List<Lotto>,
-        winningLotto: WinningLotto,
-    ): Map<Rank, Int> {
-        return LottoResult().getWinningCounts(lottoBundle, winningLotto)
-    }
-
-    private fun calculateProfit(
-        input: String,
-        winningCounts: Map<Rank, Int>,
-    ): String {
-        return LottoResult().calculateProfit(input, winningCounts)
     }
 }
